MultiLayerQG

From continuous to discrete

Under the quasi-geostrophic approximation with stratified dynamics both the horizontal flow, $u(x, y, z, t)$, $v(x, y, z, t)$, and the buoyancy $b(x, y, z, t)$ are obtained via a streamfunction $\psi(x, y, z, t)$:

\[\begin{aligned} u &= - \partial_y \psi , \\ v &= \partial_x \psi , \\ b &= f_0 \partial_z \psi . \end{aligned}\]

The dynamics are determined from the quasi-geostrophic potential vorticity (QGPV)

\[\begin{aligned} q &= \partial_x v - \partial_y u + \partial_z \left( \frac{f_0}{N^2} b \right) , \\ &= (\partial_x^2 + \partial_y^2) \psi + \partial_z \left( \frac{f_0^2}{N^2} \partial_z \psi \right) . \end{aligned}\]

where $N^2 = - \rho_0^{-1} g \, \mathrm{d} \rho / \mathrm{d}z$ is the square of the Brunt-Väisälä frequency for a Boussinesq fluid, where $\rho_0$ is a reference density and $g$ the gravitational constant.

To go from the continuous (in $z$) dynamics above to a discrete dynamics we can integrate all quantities over the height of each fluid layer. An alternative is to replace the vertical derivatives that appear both in the buoyancy definition and in the Brunt-Väisälä frequency with a discrete approximation.

We use subscripts $j$ to denote quantities that are constant within each fluid layer. For example, the heights of each fluid layer are $H_j$, $j = 1, \dots, n$, where $n$ is the number of fluid layers. Also, $\rho_j$ is the density of each layer and $\psi_j$ and $q_j$ the streamfunction and QGPV associated with that fluid layer.

The buoyancy is associated with the fluid interfaces. We use semi-integer subscript values to denote a fluid interface, e.g., $b_{5/2}$ is the buoyancy that corresponds to the interface between fluid layers 2 and 3. Thus, from the relationship that connects buoyancy $b$ and streamfunction $\psi$ we approximate the buoyancies at each interface as

\[ b_{j+1/2} = f_0 \frac{\psi_j - \psi_{j+1}}{(H_j + H_{j+1})/2} .\]

Above we divided the difference of streamfunction at the two layers with the distance between the center of the two layers, $(H_j + H_{j+1})/2$.

Similarly, since the Brunt-Väisälä frequency is proportional to the $z$-derivative of the density, it also corresponds to the fluid interfaces, i.e.,

\[ N^2_{j+1/2} = - \frac{g}{\rho_0} \frac{\rho_j - \rho_{j+1}}{(H_j + H_{j+1})/2} .\]

Combining the above we have that the vortex stretching term for the $j$-th layer is

\[\begin{aligned} \left . \partial_z \left( \frac{f_0}{N^2} b \right) \right |_{j} & = \frac{1}{H_j} \left( \frac{f_0}{N^2_{j-1/2}} b_{j-1/2} - \frac{f_0}{N^2_{j+1/2}} b_{j+1/2} \right ) \\ & = \frac{1}{H_j} \left[ \dfrac{f_0}{- \frac{g}{\rho_0} \frac{\rho_{j-1} - \rho_{j}}{(H_{j-1}+H_{j})/2}} f_0 \frac{\psi_{j-1} - \psi_{j}}{(H_{j-1}+H_{j})/2} - \dfrac{f_0}{- \frac{g}{\rho_0} \frac{\rho_j - \rho_{j+1}}{(H_j+H_{j+1})/2}} f_0 \frac{\psi_j - \psi_{j+1}}{(H_j+H_{j+1})/2} \right ] \\ & = \frac{1}{H_j} \left[\dfrac{f_0^2}{g \frac{\rho_j - \rho_{j-1} }{\rho_0}} (\psi_{j-1} - \psi_j) - \dfrac{f_0^2}{g \frac{\rho_{j+1} - \rho_j}{\rho_0}} (\psi_j - \psi_{j+1}) \right ] \\ & = \dfrac{f_0^2}{g'_{j-1/2} H_j} (\psi_{j-1} - \psi_j) - \dfrac{f_0^2}{g'_{j+1/2} H_j} (\psi_j - \psi_{j+1}) , \end{aligned}\]

where above we defined the reduced gravity that corresponds to the $j+1/2$ interface

\[ g'_{j+1/2} = g \frac{\rho_{j+1} - \rho_j}{\rho_0} .\]

Τhe the reference density $\rho_0$ appears in the denominator of the reduced gravity. One can use the mean density as the reference density, e.g.,

\[ \rho_0 = \dfrac{\sum_j \rho_j H_j}{\sum_j H_j} .\]

Basic Equations

This module solves the layered quasi-geostrophic equations on a beta plane of variable fluid depth $H - h(x, y)$. The flow in each layer is obtained through a streamfunction $\psi_j$ as $(u_j, v_j) = (-\partial_y \psi_j, \partial_x \psi_j)$, $j = 1, \dots, n$, where $n$ is the number of fluid layers.

The QGPV in each layer is

\[ \mathrm{QGPV}_j = q_j + \underbrace{f_0 + \beta y}_{\textrm{planetary PV}} + \delta_{j, n} \underbrace{\frac{f_0 h}{H_n}}_{\textrm{topographic PV}}, \quad j = 1, \dots, n .\]

where $q_j$ incorporates the relative vorticity in each layer $\nabla^2 \psi_j$ and the vortex stretching terms:

\[\begin{aligned} q_1 &= \nabla^2 \psi_1 + F_{3/2, 1} (\psi_2 - \psi_1), \\ sq_j &= \nabla^2 \psi_j + F_{j-1/2, j} (\psi_{j-1} - \psi_j) + F_{j+1/2, j} (\psi_{j+1} - \psi_j) , \quad j = 2, \dots, n-1, \\ q_n &= \nabla^2 \psi_n + F_{n-1/2, n} (\psi_{n-1} - \psi_n). \end{aligned}\]

with

\[F_{j+1/2, k} = \frac{f_0^2}{g'_{j+1/2} H_k} .\]

In view of the relationships above, when we convert to Fourier space $q$'s and $\psi$'s are related via the matrix equation

\[\begin{pmatrix} \widehat{q}_{𝐤, 1} \\ \vdots \\ \widehat{q}_{𝐤, n} \end{pmatrix} = \underbrace{\left(-|𝐤|^2 \mathbb{1} + \mathbb{F} \right)}_{\equiv \mathbb{S}_{𝐤}} \begin{pmatrix} \widehat{\psi}_{𝐤, 1} \\ \vdots \\ \widehat{\psi}_{𝐤, n} \end{pmatrix}\]

where

\[\mathbb{F} \equiv \begin{pmatrix} -F_{3/2, 1} & F_{3/2, 1} & 0 & \cdots & 0 \\ F_{3/2, 2} & -(F_{3/2, 2} + F_{5/2, 2}) & F_{5/2, 2} & & \vdots \\ 0 & \ddots & \ddots & \ddots & \\ \vdots & & & & 0 \\ 0 & \cdots & 0 & F_{n-1/2, n} & -F_{n-1/2, n} \end{pmatrix} .\]

Including an imposed zonal flow $U_j(y)$ in each layer, the equations of motion are:

\[\partial_t q_j + \mathsf{J}(\psi_j, q_j ) + (U_j - \partial_y\psi_j) \partial_x Q_j + U_j \partial_x q_j + (\partial_y Q_j)(\partial_x \psi_j) = -\delta_{j, n} \mu \nabla^2 \psi_n - \nu (-1)^{n_\nu} \nabla^{2 n_\nu} q_j ,\]

with

\[\begin{aligned} \partial_x Q_j &\equiv \delta_{j, n} \partial_x \eta . \partial_y Q_j &\equiv \beta - \partial_y^2 U_j - (1 - \delta_{j, 1}) F_{j-1/2, j} (U_{j-1} - U_j) - (1 - \delta_{j,n}) F_{j+1/2, j} (U_{j+1} - U_j) + \delta_{j, n} \partial_y \eta , \\ \end{aligned}\]

Implementation

Matrices $\mathbb{S}_{𝐤}$ as well as $\mathbb{S}^{-1}_{𝐤}$ are included in params as params.S and params.S⁻¹ respectively. Additionally, the background PV gradients $\partial_x Q$ and $\partial_y Q$ are also included in the params under params.Qx and params.Qy.

One can get the $\widehat{\psi}_j$ from $\widehat{q}_j$ via streamfunctionfrompv!(ψh, qh, params, grid), while the inverse, i.e., obtain $\widehat{q}_j$ from $\widehat{\psi}_j$, is done via pvfromstreamfunction!(qh, ψh, params, grid).

The equations of motion are time-stepped forward in Fourier space:

\[ \partial_t \widehat{q}_j = - \widehat{\mathsf{J}(\psi_j, q_j)} - \widehat{U_j \partial_x Q_j} - \widehat{U_j \partial_x q_j} + \widehat{(\partial_y \psi_j) \partial_x Q_j} - \widehat{(\partial_x \psi_j)(\partial_y Q_j)} + \delta_{j, n} \mu |𝐤|^{2} \widehat{\psi}_n - \nu |𝐤|^{2n_\nu} \widehat{q}_j .\]

In doing so the Jacobian is computed in the conservative form: $\mathsf{J}(f, g) = \partial_y [ (\partial_x f) g] - \partial_x[ (\partial_y f) g]$.

The state variable sol consists of the Fourier transforms of $q_j$ at each layer, i.e., qh.

The linear operator is constructed in Equation

GeophysicalFlows.MultiLayerQG.Equation — Function

Equation(params, grid)

Return the equation for a multi-layer quasi-geostrophic problem with params and grid. The linear opeartor $L$ includes only (hyper)-viscosity and is computed via hyperviscosity(params, grid).

The nonlinear term is computed via function calcN!.

GeophysicalFlows.MultiLayerQG.hyperviscosity — Function

hyperviscosity(params, grid)

Return the linear operator L that corresponds to (hyper)-viscosity of order $n_ν$ with coefficient $ν$ for $n$ fluid layers.

\[L_j = - ν |𝐤|^{2 n_ν}, \ j = 1, ...,n .\]

The nonlinear terms are computed via

GeophysicalFlows.MultiLayerQG.calcN! — Function

calcN!(N, sol, t, clock, vars, params, grid)

Compute the nonlinear term, that is the advection term, the bottom drag, and the forcing:

\[N_j = - \widehat{𝖩(ψ_j, q_j)} - \widehat{U_j ∂_x Q_j} - \widehat{U_j ∂_x q_j} + \widehat{(∂_y ψ_j)(∂_x Q_j)} - \widehat{(∂_x ψ_j)(∂_y Q_j)} + δ_{j, n} μ |𝐤|^2 ψ̂_n + F̂_j .\]

which in turn calls calcN_advection! and addforcing!.

Linearized MultiLayerQG dynamics

The MultiLayerQG module includes also version of the dynamics that is linearized about a base flow $U_j(y)$, $j = 1, \dots, n$; see LinearEquation, calcNlinear!, and calcN_linearadvection!.

Parameters and Variables

All required parameters are gathered in Params and all module variables are found in Vars.

For a problem without any forcing ($F = 0$), vars is constructed with DecayingVars. For the forced problems ($F \ne 0$) the vars is constructed either via ForcedVars when the forcing is deterministic or via StochasticForcedVars when the forcing is stochastic.

Helper functions

GeophysicalFlows.MultiLayerQG.set_q! — Function

set_q!(sol, params, vars, grid, q)
set_q!(prob, q)

Set the solution prob.sol as the transform of q and update variables.

GeophysicalFlows.MultiLayerQG.set_ψ! — Function

set_ψ!(params, vars, grid, sol, ψ)
set_ψ!(prob, ψ)

Set the solution prob.sol to the transform qh that corresponds to streamfunction ψ and update variables.

GeophysicalFlows.MultiLayerQG.updatevars! — Function

updatevars!(vars, params, grid, sol)
updatevars!(prob)

Update all problem variables using sol.

Diagnostics

The eddy kinetic energy in each layer and the eddy potential energy that corresponds to each fluid interface is computed via energies:

GeophysicalFlows.MultiLayerQG.energies — Function

energies(vars, params, grid, sol)
energies(prob)

Return the kinetic energy of each fluid layer KE$_1, ...,$ KE$_{n}$, and the potential energy of each fluid interface PE$_{3/2}, ...,$ PE$_{n-1/2}$, where $n$ is the number of layers in the fluid. (When $n=1$, only the kinetic energy is returned.)

The kinetic energy at the $j$-th fluid layer is

\[𝖪𝖤_j = \frac{H_j}{H} \int \frac1{2} |{\bf ∇} ψ_j|^2 \frac{𝖽x 𝖽y}{L_x L_y} = \frac1{2} \frac{H_j}{H} \sum_{𝐤} |𝐤|² |ψ̂_j|², \ j = 1, ..., n ,\]

while the potential energy that corresponds to the interface $j+1/2$ (i.e., the interface between the $j$-th and $(j+1)$-th fluid layer) is

\[𝖯𝖤_{j+1/2} = \int \frac1{2} \frac{f₀^2}{g'_{j+1/2} H} (ψ_j - ψ_{j+1})^2 \frac{𝖽x 𝖽y}{L_x L_y} = \frac1{2} \frac{f₀^2}{g'_{j+1/2} H} \sum_{𝐤} |ψ̂_j - ψ̂_{j+1}|², \ j = 1, ..., n-1 .\]

The lateral eddy fluxes in each layer and the vertical fluxes across fluid interfaces are computed via fluxes:

GeophysicalFlows.MultiLayerQG.fluxes — Function

fluxes(vars, params, grid, sol)
fluxes(prob)

Return the lateral eddy fluxes within each fluid layer, lateralfluxes$_1,...,$lateralfluxes$_n$ and also the vertical eddy fluxes at each fluid interface, verticalfluxes$_{3/2},...,$verticalfluxes$_{n-1/2}$, where $n$ is the total number of layers in the fluid. (When $n=1$, only the lateral fluxes are returned.)

The lateral eddy fluxes within the $j$-th fluid layer are

\[\textrm{lateralfluxes}_j = \frac{H_j}{H} \int U_j v_j ∂_y u_j \frac{𝖽x 𝖽y}{L_x L_y} , \ j = 1, ..., n ,\]

while the vertical eddy fluxes at the $j+1/2$-th fluid interface (i.e., interface between the $j$-th and $(j+1)$-th fluid layer) are

\[\textrm{verticalfluxes}_{j+1/2} = \int \frac{f₀²}{g'_{j+1/2} H} (U_j - U_{j+1}) \, v_{j+1} ψ_{j} \frac{𝖽x 𝖽y}{L_x L_y} , \ j = 1, ..., n-1.\]

Examples

examples/multilayerqg_2layer.jl: Simulate the growth and equilibration of baroclinic eddy turbulence in the Phillips 2-layer model.